server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Muhammad Bilal Khan <lofi.audit@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
rich = "^13.9.4"
modal = "^0.68.44"
cerebras-cloud-sdk = "^1.15.0"
diffusers = "^0.31.0"
transformers = "^4.47.1"
accelerate = "^1.2.1"
fastapi = {extras = ["standard"], version = "^0.115.6"}
torch = "^2.5.1"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
slowapi = "^0.1.9"
opencv-python = "^4.10.0.84"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
